+++ bitframe-authentication/services/server/core/src/commonMain/kotlin/bitframe/authentication/server/signin/SignInServiceConfig.kt
@@ -1,10 +1,7 @@
 package bitframe.authentication.server.signin
 
 import bitframe.authentication.users.UsersDao
-import bitframe.events.EventBus
 import bitframe.service.config.ServiceConfig
-import kotlinx.coroutines.CoroutineScope
-import kotlin.jvm.JvmOverloads
 
 interface SignInServiceConfig : ServiceConfig {
     val usersDao: UsersDao