+++ bitframe-utils/events/core/src/commonMain/kotlin/events/EventBus.kt
@@ -1,0 +1,10 @@
+@file:JsExport
+
+package events
+
+import kotlin.js.JsExport
+
+abstract class EventBus {
+    abstract fun <D> dispatch(event: Event<D>)
+    abstract fun <D> subscribe(eventId: String, callback: (D) -> Unit): Subscriber<D>
+}