+++ bitframe-utils/templater/src/commonMain/kotlin/templater/Parameter.kt
@@ -1,0 +1,7 @@
+package templater
+
+data class Parameter(
+    val raw: String
+) {
+    val name by lazy { raw.removeSurrounding(prefix = "{{", suffix = "}}").trim() }
+}