+++ pi-monitor/pi-monitor-client/browser/react/src/jsMain/kotlin/pimonitor/reakt/DropDown.kt
@@ -1,6 +1,6 @@
 package pimonitor.reakt
 
-import bitframe.presenters.fields.DropDownInputField
+import presenters.fields.DropDownInputField
 import kotlinx.css.*
 import kotlinx.css.properties.border
 import kotlinx.css.properties.boxShadow