+++ pi-monitor/pi-monitor-client/sdks/full/src/commonMain/kotlin/pimonitor/evaluation/businesses/BusinessViewModel.kt
@@ -1,6 +1,6 @@
 package pimonitor.evaluation.businesses
 
-import bitframe.presenters.collections.tableOf
+import presenters.collections.tableOf
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.collect
@@ -10,7 +10,6 @@ import later.await
 import logging.logger
 import pimonitor.evaluation.businesses.BusinessesIntent.LoadBusinesses
 import pimonitor.evaluation.businesses.BusinessesIntent.ShowBusinessForm
-import pimonitor.evaluation.businesses.BusinessesService
 import pimonitor.monitored.MonitoredBusiness
 import viewmodel.ViewModel
 import pimonitor.evaluation.businesses.BusinessesIntent as Intent
@@ -19,6 +18,7 @@ import pimonitor.evaluation.businesses.BusinessesState as State
 class BusinessViewModel(
     val service: BusinessesService
 ) : ViewModel<Intent, State>(State.Loading("Loading business")) {
+
     override fun CoroutineScope.execute(i: Intent): Any = when (i) {
         LoadBusinesses -> loadBusiness()
         ShowBusinessForm -> ui.value = State.BusinessForm()
@@ -49,6 +49,7 @@ class BusinessViewModel(
         column("NCF") { "" }
         column("V/day") { "" }
         actions("Actions") {
+            action("Invite to share reports") {}
             action("Delete") { tree.log("Deleting: ${it.data.name}") }
             action("View") { tree.log("Viewing: ${it.data.name}") }
         }