+++ pi-monitor/pi-monitor-client/sdks/full/src/jsMain/kotlin/pimonitor/evaluation/businesses/exports/BusinessesScope.kt
@@ -8,6 +8,7 @@ import pimonitor.client.evaluation.businesses.BusinessesService
 import pimonitor.evaluation.businesses.BusinessViewModel
 import pimonitor.monitored.MonitoredBusiness
 import useEventHandler
+import useViewModelState
 import viewmodel.ViewModel
 import pimonitor.evaluation.businesses.BusinessesIntent as Intent
 import pimonitor.evaluation.businesses.BusinessesState as State
@@ -22,4 +23,6 @@ class BusinessesScope(val service: PiMonitorService) {
     val useBusinessAddedEvent: (callback: (MonitoredBusiness) -> Unit) -> Unit = { callback ->
         useEventHandler(service.signIn.config.bus, BusinessesService.CREATE_BUSINESS_EVENT_ID, callback)
     }
+
+    val useStateFromViewModel = { useViewModelState(viewModel) }
 }