+++ pi-monitor/pi-monitor-server/src/test/kotlin/unit/utils/Configuration.kt
@@ -1,14 +1,10 @@
 package unit.utils
 
-import bitframe.ApplicationUnderTest
-import pimonitor.PiMonitorApplicationConfig
-import bitframe.events.InMemoryEventBus
+import events.InMemoryEventBus
 import bitframe.server.InMemoryDaoProvider
 import bitframe.service.config.ServiceConfig
 import cache.Cache
 import cache.MockCache
-import pimonitor.PiMonitorServer
-import java.io.File
 
 @JvmField
 val DAO_PROVIDER_UNDER_TEST = InMemoryDaoProvider()