+++ pi-monitor/pi-monitor-services/core/src/commonMain/kotlin/pimonitor/evaluation/businesses/BusinessesService.kt
@@ -2,7 +2,7 @@
 
 package pimonitor.evaluation.businesses
 
-import bitframe.events.Event
+import events.Event
 import bitframe.service.config.ServiceConfig
 import later.Later
 import later.await